+++ src/page/LoginAdmin/LoginAdmin.jsx
@@ -1,14 +1,14 @@
-import AuthInput from '../component/authInput/AuthInput';
-import siteLogo from '../assets/logo.svg';
-import '../styles/LoginPage.scss';
+import AuthInput from '../../component/authInput/AuthInput';
+import siteLogo from '../../assets/logo.svg';
+import '../LoginAdmin/LoginAdmin.scss';
 
-const LoginPage = () => {
+const LoginAdmin = () => {
   return (
-    <div className="loginContainer">
+    <div className="loginAdminContainer">
       <div>
         <img className="logo" src={siteLogo} alt="logo" />
       </div>
-      <h1 className="title">登入Alphitter</h1>
+      <h1 className="title">後台登入</h1>
       <div className="inputContainer">
         <AuthInput label="帳號" placeholder="請輸入帳號" value="" />
 
@@ -21,12 +21,10 @@ const LoginPage = () => {
       </div>
       <button className="btn">登入</button>
       <div className="switchSec">
-        <span className="signupSwitch">註冊</span>
-        <span className="dot"></span>
-        <span className="adminSwitch">後台登入</span>
+        <span className="adminSwitch">前台登入</span>
       </div>
     </div>
   );
 };
 
-export default LoginPage;
+export default LoginAdmin;
