+++ src/page/LoginPage/LoginPage.jsx
@@ -1,14 +1,14 @@
-import AuthInput from '../component/authInput/AuthInput';
-import siteLogo from '../assets/logo.svg';
-import '../styles/LoginAdmin.scss';
+import AuthInput from '../../component/authInput/AuthInput';
+import siteLogo from '../../assets/logo.svg';
+import '../LoginPage/LoginPage.scss';
 
-const LoginAdmin = () => {
+const LoginPage = () => {
   return (
-    <div className="loginAdminContainer">
+    <div className="loginContainer">
       <div>
         <img className="logo" src={siteLogo} alt="logo" />
       </div>
-      <h1 className="title">後台登入</h1>
+      <h1 className="title">登入Alphitter</h1>
       <div className="inputContainer">
         <AuthInput label="帳號" placeholder="請輸入帳號" value="" />
 
@@ -21,10 +21,12 @@ const LoginAdmin = () => {
       </div>
       <button className="btn">登入</button>
       <div className="switchSec">
-        <span className="adminSwitch">前台登入</span>
+        <span className="signupSwitch">註冊</span>
+        <span className="dot"></span>
+        <span className="adminSwitch">後台登入</span>
       </div>
     </div>
   );
 };
 
-export default LoginAdmin;
+export default LoginPage;
