+++ src/page/LoginPage/LoginPage.scss
@@ -22,6 +22,7 @@
     display: flex;
     flex-direction: column;
     gap: 32px;
+    width: 100%;
   }
 
   .btn {
